analiseapi/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
pydantic
tensorflow-intel
numpy
pandas
scikeras==0.13.0
joblib
